+++ src/mruby_integration/text.cpp
@@ -18,43 +18,6 @@ mrb_draw_fps(mrb_state* mrb, mrb_value) -> mrb_value
 }
 
 auto
-mrb_draw_text(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  mrb_value text;
-  mrb_int x, y, font_size;
-  Color* colour;
-  mrb_get_args(mrb, "Siiid", &text, &x, &y, &font_size, &colour, &Colour_type);
-
-  DrawText(mrb_str_to_cstr(mrb, text), x, y, font_size, *colour);
-  return mrb_nil_value();
-}
-
-auto
-mrb_draw_text_ex(mrb_state* mrb, mrb_value) -> mrb_value
-{
-  Font* font;
-  mrb_value text;
-  Vector2* position;
-  mrb_float font_size, spacing;
-  Color* colour;
-  mrb_get_args(mrb,
-               "dSdffd",
-               &font,
-               &Font_type,
-               &text,
-               &position,
-               &Vector2_type,
-               &font_size,
-               &spacing,
-               &colour,
-               &Colour_type);
-
-  DrawTextEx(
-    *font, mrb_str_to_cstr(mrb, text), *position, font_size, spacing, *colour);
-  return mrb_nil_value();
-}
-
-auto
 mrb_measure_text_ex(mrb_state* mrb, mrb_value) -> mrb_value
 {
   Font* font;
@@ -78,10 +41,6 @@ append_text(mrb_state* mrb)
 {
   mrb_define_method(
     mrb, mrb->kernel_module, "draw_fps", mrb_draw_fps, MRB_ARGS_REQ(2));
-  mrb_define_method(
-    mrb, mrb->kernel_module, "draw_text", mrb_draw_text, MRB_ARGS_REQ(5));
-  mrb_define_method(
-    mrb, mrb->kernel_module, "draw_text_ex", mrb_draw_text_ex, MRB_ARGS_REQ(6));
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "measure_text_ex",