+++ test/text_test.rb
@@ -1,30 +1,5 @@
 class Test
   class Text_Test < MTest::Unit::TestCaseWithAnalytics
-    def test_draw_text
-      skip_unless_display_present
-
-      set_window_title(__method__.to_s)
-      clear_and_draw do
-        draw_text("a", 0, 0, 10, BLACK)
-      end
-      flush_frame
-
-      assert_equal fixture_draw_text, get_screen_data.data
-    end
-
-    def test_draw_text_ex
-      skip_unless_display_present
-
-      set_window_title(__method__.to_s)
-      font = Font.new("./assets/tiny.ttf")
-      clear_and_draw do
-        draw_text_ex(font, "x", Vector2.new(0, 0), 12, 0, BLACK)
-      end
-
-      assert_within 97, fixture_draw_text_ex, get_screen_data.data
-      font.unload
-    end
-
     def test_measure_text_ex
       skip_unless_display_present